+++ packages/e2e-contract-tests/playwright/utils/index.ts
@@ -1,0 +1,1 @@
+export * from './setup';
